Create README.md
At the beginning of the pandemic, I was embarking on a PhD sandwich in Computational Modeling. However, as we know, everything changed and my trip was definitively canceled due to the blockage of the PhD scholarship. In view of this, I gave up my apartment lease in the city. My daughter, my dog, and I went to a beach house that had been abandoned by the family. We did a little renovation and started a life far away from technology. After the renovation was finished, I was already raising chickens in the house and already with a new romance, I was called to take a public exam to work as a pharmacist in the public health system in another state approximately 2,000 km away. So, we all left the beach house and came to where I am now. The romance ended, but now I am starting a new journey. In a precarious work system that is eager for simple technologies and innovations. There is a lot to do here, after all, we are neither in the Stone Age nor the Saw Metal Age.